prod/api/requirements.txt
fastapi
uvicorn[standard]
pillow
ultralytics
python-multipart
